move clientIP to http.go
