+++ http.go
@@ -149,3 +149,30 @@ func (le *DefaultHTTPLogEntry) isIgnored() bool {
 	}
 	return false
 }
+
+var (
+	trueClientIP          = http.CanonicalHeaderKey("True-Client-IP")
+	xForwardedFor         = http.CanonicalHeaderKey("X-Forwarded-For")
+	xRealIP               = http.CanonicalHeaderKey("X-Real-IP")
+	xEnvoyExternalAddress = http.CanonicalHeaderKey("X-Envoy-External-Address")
+)
+
+// clientIP returns the IP of the client.
+// If a header identifying the real IP exists, the value of the header will be used.
+func clientIP(h http.Header) string {
+	if tcip := h.Get(trueClientIP); tcip != "" {
+		return tcip
+	} else if xrip := h.Get(xRealIP); xrip != "" {
+		return xrip
+	} else if xff := h.Get(xForwardedFor); xff != "" {
+		i := strings.Index(xff, ",")
+		if i == -1 {
+			i = len(xff)
+		}
+		return xff[:i]
+	} else if xeea := h.Get(xEnvoyExternalAddress); xeea != "" {
+		return xeea
+	}
+
+	return ""
+}
