+++ http_option_test.go
@@ -1,7 +1,6 @@
 package accesslog
 
 import (
-	"net/http"
 	"reflect"
 	"testing"
 )
@@ -59,50 +58,6 @@ func Test_headerMap(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := headerMap(tt.hs); !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("headerMap() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_clientIP(t *testing.T) {
-	tests := []struct {
-		name string
-		h    http.Header
-		want string
-	}{
-		{
-			name: "true-client-ip",
-			h: http.Header{
-				"True-Client-Ip": []string{"255.255.255.255"},
-			},
-			want: "255.255.255.255",
-		},
-		{
-			name: "x-forwarded-for",
-			h: http.Header{
-				"X-Forwarded-For": []string{"255.255.255.255"},
-			},
-			want: "255.255.255.255",
-		},
-		{
-			name: "x-real-ip",
-			h: http.Header{
-				"X-Real-Ip": []string{"255.255.255.255"},
-			},
-			want: "255.255.255.255",
-		},
-		{
-			name: "x-envoy-external-address",
-			h: http.Header{
-				"X-Envoy-External-Address": []string{"255.255.255.255"},
-			},
-			want: "255.255.255.255",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := clientIP(tt.h); got != tt.want {
-				t.Errorf("clientIP() = %v, want %v", got, tt.want)
 			}
 		})
 	}
